requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.0.1
